Add readme for filtering template issues script
